Implemented calculator controller with TDD

+++ TestDrivenDevelopmentDemo.Api/Calculator.cs
@@ -6,24 +6,24 @@ namespace TestDrivenDevelopmentDemo.Api
 {
     public class Calculator
     {
-        public int Add(int value1, int value2)
+        public double Add(double value1, double value2)
         {
             return value1 + value2;
         }
 
-        public int Subtract(int value1, int value2)
+        public double Subtract(double value1, double value2)
         {
             return value1 - value2;
         }
 
-        public int Multiply(int value1, int value2)
+        public double Multiply(double value1, double value2)
         {
             return value1 * value2;
         }
 
-        public int Divide(int value1, int value2)
+        public double Divide(double value1, double value2)
         {
-            if (value2 == 0)
+            if (value2 == 0.0)
                 throw new InvalidOperationException("Argument cannot be zero");
 
             return value1 / value2;
